Viewer refactor - simplify a bit
