+++ src/viewer/assets/lib/line.js
@@ -1,6 +1,10 @@
 import {
     createAxisLabelFormatter,
 } from './units.js';
+import {
+    getBaseOption,
+    getBaseYAxisOption,
+} from './charts/base.js';
 
 /**
  * Creates a line chart configuration for ECharts
@@ -9,11 +13,13 @@ import {
  * @param {Object} plotSpec - Plot specification with data and options
  * @returns {Object} ECharts configuration object
  */
-export function createLineChartOption(baseOption, plotSpec) {
+export function createLineChartOption(plotSpec) {
     const {
         data,
         opts
     } = plotSpec;
+
+    const baseOption = getBaseOption(opts.title);
 
     if (!data || data.length < 2) {
         return baseOption;
@@ -32,40 +38,9 @@ export function createLineChartOption(baseOption, plotSpec) {
     const minValue = format.min;
     const maxValue = format.max;
 
-    const yAxis = {
-        type: logScale ? 'log' : 'value',
-        logBase: 10,
-        scale: true,
-        min: minValue,
-        max: maxValue,
-        axisLine: {
-            lineStyle: {
-                color: '#ABABAB'
-            }
-        },
-        axisLabel: {
-            color: '#ABABAB',
-            margin: 16, // Fixed consistent margin for all charts
-            formatter: unitSystem ?
-                createAxisLabelFormatter(unitSystem) :
-                function (value) {
-                    // Use scientific notation for large/small numbers
-                    if (Math.abs(value) > 10000 || (Math.abs(value) > 0 && Math.abs(value) < 0.01)) {
-                        return value.toExponential(1);
-                    }
-                    return value;
-                }
-        },
-        splitLine: {
-            lineStyle: {
-                color: 'rgba(171, 171, 171, 0.2)'
-            }
-        }
-    };
-
     return {
         ...baseOption,
-        yAxis,
+        yAxis: getBaseYAxisOption(logScale, minValue, maxValue, unitSystem),
         tooltip: {
             ...baseOption.tooltip,
             valueFormatter: unitSystem ?
