+++ src/viewer/assets/lib/scatter.js
@@ -3,6 +3,10 @@
 import {
     createAxisLabelFormatter,
 } from './units.js';
+import {
+    getBaseOption,
+    getBaseYAxisOption,
+} from './charts/base.js';
 
 /**
  * Creates a scatter chart configuration for ECharts
@@ -11,11 +15,13 @@ import {
  * @param {Object} plotSpec - Plot specification with data and options
  * @returns {Object} ECharts configuration object
  */
-export function createScatterChartOption(baseOption, plotSpec) {
+export function createScatterChartOption(plotSpec) {
     const {
         data,
         opts
     } = plotSpec;
+
+    const baseOption = getBaseOption(opts.title);
 
     if (!data || data.length < 2) {
         return baseOption;
@@ -73,45 +79,10 @@ export function createScatterChartOption(baseOption, plotSpec) {
         unitSystem === 'time';
     // TODO: remove the above second-guessing and just use the unit system.
 
-    const yAxis = {
-        type: logScale ? 'log' : 'value',
-        logBase: 10,
-        scale: true,
-        min: minValue,
-        max: maxValue,
-        axisLine: {
-            lineStyle: {
-                color: '#ABABAB'
-            }
-        },
-        axisLabel: {
-            color: '#ABABAB',
-            margin: 12, // Fixed consistent margin for all charts
-            formatter: unitSystem ?
-                createAxisLabelFormatter(unitSystem) :
-                function (value) {
-                    // Format log scale labels more compactly if needed
-                    if (logScale && Math.abs(value) >= 1000) {
-                        return value.toExponential(0);
-                    }
-                    // Use scientific notation for large/small numbers
-                    if (Math.abs(value) > 10000 || (Math.abs(value) > 0 && Math.abs(value) < 0.01)) {
-                        return value.toExponential(1);
-                    }
-                    return value;
-                }
-        },
-        splitLine: {
-            lineStyle: {
-                color: 'rgba(171, 171, 171, 0.2)'
-            }
-        }
-    };
-
     // Return scatter chart configuration with reliable time axis
     return {
         ...baseOption,
-        yAxis,
+        yAxis: getBaseYAxisOption(logScale, minValue, maxValue, unitSystem),
         tooltip: {
             ...baseOption.tooltip,
             valueFormatter: unitSystem ?
